SV-609 Add LoadViewModel implementation to CertificateCheckController
