+++ src/SFA.DAS.AssessorService.Web/ViewModels/Certificate/CertificateBaseViewModel.cs
@@ -1,8 +1,6 @@
 ﻿using System;
-using System.Linq;
 using Newtonsoft.Json;
 using SFA.DAS.AssessorService.Domain.JsonData;
-using SFA.DAS.AssessorService.Web.Controllers;
 
 namespace SFA.DAS.AssessorService.Web.ViewModels.Certificate
 {
@@ -25,13 +23,6 @@ namespace SFA.DAS.AssessorService.Web.ViewModels.Certificate
             StandardUId = cert.StandardUId;            
         }
 
-        public void SetStandardHasVersionsAndOptions(CertificateSession certSession)
-        {
-            StandardHasOptions = certSession.Options != null && certSession.Options.Any();
-            StandardHasSingleOption = certSession.Options == null || certSession.Options.Count <= 1;
-            StandardHasSingleVersion = certSession.Versions == null || certSession.Versions.Count <= 1;
-        }
-
         public Guid Id { get; set; }
         public string FamilyName { get; set; }
         public string GivenNames { get; set; }
@@ -41,9 +32,6 @@ namespace SFA.DAS.AssessorService.Web.ViewModels.Certificate
         public string StandardUId { get; set; }
         public string Uln { get; set; }
         public int Level { get; set; }
-        public bool StandardHasSingleOption { get; set; }
-        public bool StandardHasOptions { get; set; }
-        public bool StandardHasSingleVersion { get; set; }
         public bool BackToCheckPage { get; set; }
     }
 }
